Fixed bug with Fraction.GetHashcode(). Made the blog more GMaily.
git-svn-id: https://ironscheme.svn.codeplex.com/svn@16363 2327b42d-5241-43d6-9e2a-de5ac946f064

+++ IronScheme/IronScheme/Runtime/BigFraction.cs
@@ -472,7 +472,7 @@ namespace IronScheme.Runtime
 
 		public override Int32 GetHashCode()
 		{
-			return (Int32)this;
+			return (Int32)(double)this;
 		}
 
     public static Boolean operator <=(Fraction fraction1, Fraction fraction2)
